fix: move TLOS_SYMBOL assignment to config
